Completed go implementation

+++ go_gridGamesAi/c_makeCallback.go
@@ -2,6 +2,7 @@ package main
 
 /*
 #include <stdlib.h>
+#include <stdio.h>
 typedef float (*Score) (int*);
 float makeMyCallback(Score f, int* arr_grid){
 	return f(arr_grid);
@@ -9,7 +10,7 @@ float makeMyCallback(Score f, int* arr_grid){
 */
 import "C"
 
-func goCallback(score C.Score, grid *C.int) float64 {
+func goCallback(score C.Score, grid *C.int) float32 {
 	r := C.makeMyCallback(score, grid)
-	return float64(r)
+	return float32(r)
 }
